Rename `precompiles` to `handlers`
